Drastische Reduzierung der Klassenanzahl von 25 auf 5
