--- src/NotizProjekt_All/GUI_NotizBearbeiten.java
@@ -1,378 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package NotizProjekt_All;
-
-import NotizProjekt_All.GUI_MenuTabelle;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.ArrayList;
-import javax.swing.JOptionPane;
-
-/**
- *
- * @author tiana.trajkovska
- */
-public final class GUI_NotizBearbeiten extends javax.swing.JFrame {
-   
-    private static int NutzerID;
-    int Nutzer= GUI_Anmelden.NutzerID;
-    static int Notiznr=GUI_MenuTabelle.Notiznr;
-    private DBVerbindung konnektor;
-    private ArrayList<Notiz> notizenliste;
-   
-    
-    String titel;
-    String tag;
-    String txnotiz; 
-
-    /**
-     * Creates new form GUI_NeueNotiz
-     */
-    public GUI_NotizBearbeiten() {
-     
-    try{
-        initComponents();
-              this.konnektor = new DBVerbindung("localhost", "its-projekt", "root", "");
-            this.konnektor.open();
-        }
-         catch (ClassNotFoundException | SQLException ex) {
-            JOptionPane.showMessageDialog(rootPane, "Fehler! Verbindung konnte nicht aufgebaut werden: " + ex);
-        }
-           
-        if(GUI_MenuTabelle.OF==true){
-                 getSONotiz();
-            zeigeNotiz(); 
-           
-        }else{
-      
-        getSNotiz();
-            zeigeNotiz();
-        }
-
-     
-    
-    
-    
-        
-        
-    }
-    
-    public void getSNotiz() {
-        
-        
-        
-        
-        try {
-            ResultSet ergebnis = this.konnektor.fuehreAbfrageAus("SELECT Titel,Tag,Inhalt  from notiz where N_ID="+Notiznr+"");
-            
-            while (ergebnis.next()) {
-         
-                   titel=ergebnis.getString("Titel");
-                        tag=ergebnis.getString("Tag");
-                        txnotiz=ergebnis.getString("Inhalt");
-            }
-
-        } catch (SQLException ex) {
-            JOptionPane.showMessageDialog(rootPane, "Fehler bei der Abfrage der Datenbank: " + ex);
-    }
-        
-    } 
-    
-    public void getSONotiz() {
-        
-        
-        
-        System.out.println("getOnitz"+Notiznr);
-        try {
-            ResultSet ergebnis = this.konnektor.fuehreAbfrageAus("SELECT Titel,Tag,Inhalt from geteilte_notizen ");
-           
-            while (ergebnis.next()) {
-     
-             
-                        titel=ergebnis.getString("Titel");
-                        tag=ergebnis.getString("Tag");
-                        txnotiz=ergebnis.getString("Inhalt");
-                        
-                 }
-                
-            
-
-        } catch (SQLException ex) {
-            JOptionPane.showMessageDialog(rootPane, "Fehler bei der Abfrage der Datenbank: " + ex);
-        }
-        
-    } 
-    
-    public void zeigeNotiz(){
-       txfTitel.setText(titel );
-     txfTag.setText(tag );
-     txfNotiz.setText(txnotiz );
-        
- }
-    public void UpdateNotiz(){
-    
-    titel="";
-    tag="";
-    txnotiz="";
-    titel = txfTitel.getText();
-    tag = txfTag.getText();
-    txnotiz = txfNotiz.getText();
-        
-        
-    try{
-    
-    int ergebnis= this.konnektor.fuehreVeraenderungAus(     "UPDATE notiz SET Titel = '" + titel + "', Tag = '" + tag + "', Inhalt = '" + txnotiz + "' WHERE N_id = " + Notiznr);
-    
-    
-    } catch(SQLException ex) {
-            JOptionPane.showMessageDialog(rootPane, "Fehler! Verbindung konnte nicht aufgebaut werden: " + ex);
-        }
-    }
-    
-    
-   public void UpdateOeffentlicheNotiz(){
-    
-   
-    titel="";
-    tag="";
-    txnotiz="";
-    titel = txfTitel.getText();
-    tag = txfTag.getText();
-    txnotiz = txfNotiz.getText();
-        
-
-        
-    try{
-    int ergebnis= this.konnektor.fuehreVeraenderungAus(     "UPDATE geteilte_notizen SET Titel = '" + titel + "', Tag = '" + tag + "', Inhalt = '" + txnotiz + "");
-    }
-    
-     catch(SQLException ex) {
-            JOptionPane.showMessageDialog(rootPane, "Fehler! Verbindung konnte nicht aufgebaut werden: " + ex);
-        }
-    
-     
-    
-  
-
-     }
-    
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        cbOeffentlich = new javax.swing.JCheckBox();
-        jPanel1 = new javax.swing.JPanel();
-        jLabel1 = new javax.swing.JLabel();
-        txfTitel = new javax.swing.JTextField();
-        txfTag = new javax.swing.JTextField();
-        btnErstellen = new javax.swing.JButton();
-        btnZurueck = new javax.swing.JButton();
-        jScrollPane1 = new javax.swing.JScrollPane();
-        txfNotiz = new javax.swing.JTextPane();
-
-        cbOeffentlich.setFont(new java.awt.Font("Tahoma", 0, 14)); // NOI18N
-        cbOeffentlich.setText("Öffentlich");
-        cbOeffentlich.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                cbOeffentlichActionPerformed(evt);
-            }
-        });
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-        setBackground(new java.awt.Color(96, 96, 96));
-
-        jPanel1.setBackground(new java.awt.Color(96, 96, 96));
-
-        jLabel1.setFont(new java.awt.Font("Arial", 3, 24)); // NOI18N
-        jLabel1.setForeground(new java.awt.Color(255, 255, 255));
-        jLabel1.setText("Notiz Bearbeiten");
-
-        txfTitel.setBackground(new java.awt.Color(96, 96, 96));
-        txfTitel.setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
-        txfTitel.setForeground(new java.awt.Color(255, 255, 255));
-        txfTitel.setBorder(javax.swing.BorderFactory.createTitledBorder(null, "Titel:", javax.swing.border.TitledBorder.DEFAULT_JUSTIFICATION, javax.swing.border.TitledBorder.DEFAULT_POSITION, new java.awt.Font("Arial", 0, 14), new java.awt.Color(255, 255, 255))); // NOI18N
-        txfTitel.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                txfTitelActionPerformed(evt);
-            }
-        });
-
-        txfTag.setBackground(new java.awt.Color(96, 96, 96));
-        txfTag.setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
-        txfTag.setForeground(new java.awt.Color(255, 255, 255));
-        txfTag.setBorder(javax.swing.BorderFactory.createTitledBorder(null, "Tag:", javax.swing.border.TitledBorder.DEFAULT_JUSTIFICATION, javax.swing.border.TitledBorder.DEFAULT_POSITION, new java.awt.Font("Arial", 0, 14), new java.awt.Color(255, 255, 255))); // NOI18N
-        txfTag.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                txfTagActionPerformed(evt);
-            }
-        });
-
-        btnErstellen.setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
-        btnErstellen.setText("Übernehmen");
-        btnErstellen.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btnErstellenActionPerformed(evt);
-            }
-        });
-
-        btnZurueck.setFont(new java.awt.Font("Arial", 0, 14)); // NOI18N
-        btnZurueck.setText("Zurück");
-        btnZurueck.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btnZurueckActionPerformed(evt);
-            }
-        });
-
-        txfNotiz.setBackground(new java.awt.Color(96, 96, 96));
-        txfNotiz.setBorder(javax.swing.BorderFactory.createTitledBorder(null, "Notiz:", javax.swing.border.TitledBorder.DEFAULT_JUSTIFICATION, javax.swing.border.TitledBorder.DEFAULT_POSITION, new java.awt.Font("Arial", 0, 14), new java.awt.Color(255, 255, 255))); // NOI18N
-        txfNotiz.setForeground(new java.awt.Color(255, 255, 255));
-        jScrollPane1.setViewportView(txfNotiz);
-
-        javax.swing.GroupLayout jPanel1Layout = new javax.swing.GroupLayout(jPanel1);
-        jPanel1.setLayout(jPanel1Layout);
-        jPanel1Layout.setHorizontalGroup(
-            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel1Layout.createSequentialGroup()
-                .addGap(190, 190, 190)
-                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
-                    .addComponent(txfTag, javax.swing.GroupLayout.PREFERRED_SIZE, 200, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(txfTitel, javax.swing.GroupLayout.PREFERRED_SIZE, 200, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel1Layout.createSequentialGroup()
-                .addContainerGap(110, Short.MAX_VALUE)
-                .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 362, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(109, 109, 109))
-            .addGroup(jPanel1Layout.createSequentialGroup()
-                .addGap(41, 41, 41)
-                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(jPanel1Layout.createSequentialGroup()
-                        .addComponent(jLabel1)
-                        .addGap(20, 353, Short.MAX_VALUE))
-                    .addGroup(jPanel1Layout.createSequentialGroup()
-                        .addComponent(btnZurueck, javax.swing.GroupLayout.PREFERRED_SIZE, 87, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                        .addComponent(btnErstellen)
-                        .addGap(43, 43, 43))))
-        );
-        jPanel1Layout.setVerticalGroup(
-            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(jPanel1Layout.createSequentialGroup()
-                .addGap(27, 27, 27)
-                .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 46, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(37, 37, 37)
-                .addComponent(txfTitel, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(15, 15, 15)
-                .addComponent(txfTag, javax.swing.GroupLayout.PREFERRED_SIZE, 45, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(18, 18, 18)
-                .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 321, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 33, Short.MAX_VALUE)
-                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(btnErstellen, javax.swing.GroupLayout.PREFERRED_SIZE, 31, javax.swing.GroupLayout.PREFERRED_SIZE)
-                    .addComponent(btnZurueck, javax.swing.GroupLayout.PREFERRED_SIZE, 31, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addGap(46, 46, 46))
-        );
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(jPanel1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(jPanel1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-        );
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-
-    private void cbOeffentlichActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_cbOeffentlichActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_cbOeffentlichActionPerformed
-
-    private void txfTitelActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_txfTitelActionPerformed
-
-    }//GEN-LAST:event_txfTitelActionPerformed
-
-    private void txfTagActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_txfTagActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_txfTagActionPerformed
-
-    private void btnErstellenActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnErstellenActionPerformed
-    
-
-        titel = txfTitel.getText();
-        tag = txfTag.getText();
-        txnotiz = txfNotiz.getText();
-
-           
-
-        new GUI_MenuTabelle().setVisible(true);
-        this.dispose();
-
-    }//GEN-LAST:event_btnErstellenActionPerformed
-
-    private void btnZurueckActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnZurueckActionPerformed
-        new GUI_MenuTabelle().setVisible(true);
-        this.dispose();
-    }//GEN-LAST:event_btnZurueckActionPerformed
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(GUI_NotizBearbeiten.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(GUI_NotizBearbeiten.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(GUI_NotizBearbeiten.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(GUI_NotizBearbeiten.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        //</editor-fold>
-        //</editor-fold>
-
-        /* Create and display the form */
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                new GUI_NotizBearbeiten().setVisible(true);
-                
-            }
-        });
-    }
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JButton btnErstellen;
-    private javax.swing.JButton btnZurueck;
-    private javax.swing.JCheckBox cbOeffentlich;
-    private javax.swing.JLabel jLabel1;
-    private javax.swing.JPanel jPanel1;
-    private javax.swing.JScrollPane jScrollPane1;
-    private javax.swing.JTextPane txfNotiz;
-    private javax.swing.JTextField txfTag;
-    private javax.swing.JTextField txfTitel;
-    // End of variables declaration//GEN-END:variables
-
-    
-}
